feat: mayor changes, almost finished

+++ src/main/kotlin/com/example/iub/repositories/AeronaveRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface AeronaveRepository : JpaRepository<Aeronave, Long>
+interface AeronaveRepository : JpaRepository<Aeronave, Int>
