+++ src/main/kotlin/com/example/iub/repositories/UsuarioRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface UsuarioRepository : JpaRepository<Usuario, Long> {
+interface UsuarioRepository : JpaRepository<Usuario, Int> {
     fun findByEmail(email: String): Usuario?
 }
